Refactored BuildSeleniumRunner into a class
